PEPPER-1145-playwright-test
