+++ playwright-e2e/tests/dsm/miscellaneous/participant-withdrawal.spec.ts
@@ -12,8 +12,7 @@ import { assertTableHeaders } from 'utils/assertion-helper';
 import { logParticipantWithdrew } from 'utils/log-utils';
 import * as user from 'data/fake-user.json';
 
-// Enable after bug fix https://broadworkbench.atlassian.net/browse/PEPPER-1112
-test.describe.skip('Participants Withdrawal', () => {
+test.describe('Participants Withdrawal', () => {
   const studies = [StudyEnum.LMS];
 
     for (const study of studies) {
